requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
babel==2.17.0
beautifulsoup4==4.13.4
bs4==0.0.2
cachetools==5.5.2
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
courlan==1.3.2
cssselect==1.3.0
dateparser==1.2.1
faiss-cpu==1.11.0
filelock==3.18.0
fsspec==2025.5.1
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.1.5
htmldate==1.9.3
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jusText==3.0.2
langchain==0.3.26
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langsmith==0.4.1
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
nose==1.3.7
numpy==2.3.1
orjson==3.10.18
packaging==24.2
pillow==11.2.1
playwright==1.52.0
playwright-stealth==2.0.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
readability-lxml==0.8.4.1
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==4.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sympy==1.14.0
tenacity==9.1.2
textsplit==0.5
threadpoolctl==3.6.0
tld==0.13.1
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
zstandard==0.23.0
rank_bm25==0.2.2
